00-Vision-transformer-ViT/requirements.txt
torch>=1.7.0 
timm>=0.3.2 
Pillow>=8.0.0 
matplotlib>=3.3.0 
numpy>=1.19.0

